add express.Router support
