+++ README.md
@@ -8,54 +8,40 @@
 npm run dev
 ```
 
-Для тестирования API можно использовать Postman или curl. Ниже примеры запросов на curl:
+Для тестирования API можно использовать Postman или curl. Ниже примеры запросов:
 
 ### Авторизация пользователя
 
 ```bash
-curl -X POST http://localhost:3000/api/user/login
+POST /api/user/login
 ```
 
 ### Получить все книги
 
 ```bash
-curl http://localhost:3000/api/books
+GET /api/books
 ```
 
 ### Получить книгу по ID
 
 ```bash
-curl http://localhost:3000/api/books/1
+GET /api/books/:id
 ```
 
 ### Создать новую книгу
 
 ```bash
-curl -X POST -H "Content-Type: application/json" -d '{
-  "title": "Новая книга",
-  "description": "Описание новой книги",
-  "authors": "Автор книги",
-  "favorite": "false",
-  "fileCover": "new-book.jpg",
-  "fileName": "new-book.pdf"
-}' http://localhost:3000/api/books
+POST /api/books
 ```
 
 ### Редактировать книгу по ID
 
 ```bash
-curl -X PUT -H "Content-Type: application/json" -d '{
-  "title": "Обновлённое название",
-  "description": "Обновлённое описание",
-  "authors": "Новый автор",
-  "favorite": "true",
-  "fileCover": "updated-cover.jpg",
-  "fileName": "updated-file.pdf"
-}' http://localhost:3000/api/books/1
+PUT /api/books/:id
 ```
 
 ### Удалить книгу по ID
 
 ```bash
-curl -X DELETE http://localhost:3000/api/books/1
+DELETE /api/books/:id
 ```
