+++ src/routes/user.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+// Авторизация пользователя
+router.post('/login', (_, res) => {
+  res.status(201).json({ id: 1, mail: 'test@mail.ru' });
+});
+
+module.exports = router;
